--- ThiagoTrabalhoJava/src/controle/manutencao.java
@@ -1,14 +1,0 @@
-package controle;
-
-public class manutencao extends MenuPrincipal.Gastos {
-    private String tipo;
-
-    public String getTipo() {
-        return tipo;
-    }
-
-    public void setTipo(String tipo) {
-        this.tipo = tipo;
-    }
-    
-}
